Created the interface file for the dessert data
