+++ src/app/components/product-card/product-card.component.html
@@ -1,0 +1,1 @@
+<p>product-card works!</p>
