+++ src/app/components/product-card/product-card.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-product-card',
+  imports: [],
+  templateUrl: './product-card.component.html',
+  styleUrl: './product-card.component.scss'
+})
+export class ProductCardComponent {
+
+}
